New commit

+++ main.py
@@ -6,7 +6,7 @@ def AnaliseDeEntrada():
     st.title("Anlise de sequencia Numérica")
     stringNumeros = st.text_input('Entre com uma sequencia de Numeros sepadas por espaço: ',placeholder="1 2 34 2 5 1 2") 
     if stringNumeros:
-        numeros = np.array(stringNumeros.split(" "),dtype=int)
+        numeros = np.array(stringNumeros.split(" "),dtype=float)
         st.markdown(f"""
                     A sequencia de números : {numeros}  
                     Soma: {numeros.sum()}  
@@ -15,7 +15,7 @@ def AnaliseDeEntrada():
                     Desvio padrão: {np.std(numeros)}  
                     Valor Max: {numeros.max()}  
                     Valor Min: {numeros.min()} 
-                
+
                     """)
 def TCRS_Calculo():
     st.title("Taxa de Lixo")
